Switch to alternate darkTheme hook implementation
