+++ components/src/themed/ThemedBadge.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { Badge, useTheme } from 'react-native-paper';
-import { $DeepPartial } from '@callstack/react-theme-provider';
-import { alternateDarkTheme } from './shared/alternateDarkTheme';
+import { Badge } from 'react-native-paper';
+import { useAltDarkTheme } from './hooks/useAltDarkTheme';
 
 export type ThemedBadgeProps = React.ComponentProps<typeof Badge>;
 
@@ -14,12 +13,8 @@ export type ThemedBadgeProps = React.ComponentProps<typeof Badge>;
  */
 export const ThemedBadge: React.FC<ThemedBadgeProps> = (props) => {
     const { theme: themeOverride, ...other } = props;
-    const theme = useTheme(themeOverride);
+    const theme = useAltDarkTheme(themeOverride);
 
-    const altDarkTheme: $DeepPartial<ReactNativePaper.Theme> = Object.assign({
-        ...alternateDarkTheme(theme)
-    }, themeOverride);
-
-    return <Badge {...other} theme={theme.dark ? altDarkTheme : themeOverride} />;
+    return <Badge {...other} theme={theme} />;
 
 };
